create abstract moving_thing
