+++ player.py
@@ -1,9 +1,9 @@
-from thing import Thing
+from moving_thing import MovingThing
 from item import Item
 import pygame
 
 
-class Player(Thing):
+class Player(MovingThing):
     def __init__(self, room, x=0, y=0):
         super().__init__(room, x, y)
         self.canvas.fill((255, 0, 0))
@@ -13,9 +13,6 @@ class Player(Thing):
             pygame.K_UP:    (0, -1),
             pygame.K_DOWN:  (0, 1),
         }
-
-        self.speed = 1
-        self.vision = 10
 
     def can_move_to(self, x, y):
         return 0 <= x < self.room.width and 0 <= y < self.room.height and \
